chore(structure): split maintain/dir_graph god files; tighten grown-function exceptions
maintain.rs 2556->2101 (three inline test mods to sibling files, existing
pattern); dir_graph/mod.rs 2513->2193 (columnar-rebuild pipeline to a
private sibling — a first cut moving pub fns was discarded because the
API-profile capture groups by impl block and would have diffed with
byte-identical signatures); rollback fingerprint factored per section
(exception no longer needed); load_disk_dir and add_nodes cold prologues
extracted with their exceptions hand-tightened (338/60->217/39,
275/27->233/19). No ceilings raised, no baselines refreshed; bench hold
check: schema-shell 6.0-6.5 us, append 83-84 us, mass-SET 226-233 ns/row
— all at or below their program records.
