+++ crates/kglite/src/graph/io/file.rs
@@ -1964,6 +1964,88 @@ fn load_disk_dir(dir: &std::path::Path) -> io::Result<Arc<DirGraph>> {
             .insert(node_type.clone(), std::sync::Arc::new(schema));
     }
 
+    load_disk_column_stores(dir, &mut graph)?;
+
+    // No sync: the stores were installed straight onto the backend, which is
+    // their only owner (D1 Phase 3 deleted the DirGraph↔DiskGraph mirror).
+
+    // Load id_indices from disk.
+    //
+    // Three formats, in priority order:
+    //   1. id_indices.bin   — 0.8.28+ raw mmap-resident layout (lazy reads,
+    //      ~ms load even at Wikidata scale).
+    //   2. id_indices.bin.zst with KGLIIDX1 magic — 0.8.13 flat-CSR format
+    //      (eager decompress + HashMap rebuild; retained data fallback).
+    // Pre-0.14 bincode caches are ignored and rebuilt lazily.
+    let t = stage_timer();
+    if crate::graph::storage::GraphRead::is_disk(&graph.graph) {
+        if let Some(base) =
+            crate::graph::storage::disk::id_index::IdIndexBase::load_from(dir, &graph.interner)?
+        {
+            graph.id_indices = crate::graph::storage::disk::id_index::IdIndexStore::from_base(base);
+        } else {
+            let id_indices_path = dir.join("id_indices.bin.zst");
+            if id_indices_path.exists() {
+                if let Ok(compressed) = std::fs::read(&id_indices_path) {
+                    if let Ok(bytes) = zstd_decompress(&compressed) {
+                        if let Ok(Some(indices)) = read_id_indices_bin(&bytes, &graph.interner) {
+                            graph.id_indices.replace_with(indices);
+                        }
+                    }
+                }
+            }
+        }
+    }
+    log_stage("id_indices_load", t);
+
+    // 0.8.28+: `type_connectivity_cache` is populated lazily on first
+    // access (in `introspection/describe.rs`'s
+    // `compute_type_connectivity` fallback). Pre-loading it eagerly was
+    // costing 15+ s on slice-built graphs (128 M triples × 3 String
+    // allocations each) for data that most query workloads never touch.
+    // Read sites that miss the cache already degrade gracefully to a
+    // bounded edge scan.
+    //
+    // Opt-in eager load: `KGLITE_EAGER_TYPE_CONNECTIVITY=1`. Users that
+    // call `describe()` immediately after load can set this to amortize
+    // the cost into load instead of the first describe().
+    let t = stage_timer();
+    if std::env::var_os("KGLITE_EAGER_TYPE_CONNECTIVITY").is_some()
+        && !graph.has_type_connectivity_cache()
+    {
+        if let Ok(Some(triples)) = read_type_connectivity_bin(dir, &graph) {
+            if !triples.is_empty() {
+                *graph.type_connectivity_cache.write().unwrap() = Some(triples);
+            }
+        }
+    }
+    log_stage("type_connectivity_load", t);
+
+    load_disk_sidecars(dir, &mut graph)?;
+
+    // Backfill the connection_types O(1)-lookup cache from the loaded
+    // metadata. The v3 / file loader does this at line 1606 of read_v3;
+    // the disk loader was the only path that left it empty and relied
+    // on `has_connection_type`'s metadata-fallback branch. The fallback
+    // is correct on a freshly-loaded graph but flips into the wrong
+    // branch the moment any code path calls `register_connection_type`
+    // (which inserts into the cache and trips the "use cache" fast
+    // path on subsequent lookups). Backfilling here keeps the cache
+    // authoritative throughout the lifetime of the loaded graph.
+    graph.build_connection_types_cache();
+
+    log_stage("load_disk_dir_total", _load_t);
+
+    Ok(Arc::new(graph))
+}
+
+/// Install a disk graph's column stores — the mmap-backed `columns.bin` +
+/// `columns_meta` pair when present, otherwise the per-type
+/// `columns/<type>/columns.zst` sidecars. Split out of `load_disk_dir` to keep
+/// it under the function-complexity ceiling; cold load-time path.
+fn load_disk_column_stores(dir: &std::path::Path, graph: &mut DirGraph) -> io::Result<()> {
+    use crate::graph::io::load_timing::{log_stage, stage_timer};
+
     // Load column stores — prefer mmap-backed (columns.bin + columns_meta).
     // 0.8.12 phase-1: PR1 phase 4 moved these files to `seg_000/`. Check
     // both locations so post-phase-4 saves still take the fast mmap path
@@ -2045,68 +2127,19 @@ fn load_disk_dir(dir: &std::path::Path) -> io::Result<Arc<DirGraph>> {
         // emits `columns/<type>/columns.zst` only for types NOT in
         // `columns_meta`, so the two paths don't clash — but we still
         // check before overwriting out of caution.
-        load_column_sidecars(dir, &mut graph)?;
+        load_column_sidecars(dir, graph)?;
     } else {
         // Per-type sidecar path: load columns/<type>/columns.zst files.
-        load_column_sidecars(dir, &mut graph)?;
+        load_column_sidecars(dir, graph)?;
     }
     log_stage("column_stores_load", t);
+    Ok(())
+}
 
-    // No sync: the stores were installed straight onto the backend, which is
-    // their only owner (D1 Phase 3 deleted the DirGraph↔DiskGraph mirror).
-
-    // Load id_indices from disk.
-    //
-    // Three formats, in priority order:
-    //   1. id_indices.bin   — 0.8.28+ raw mmap-resident layout (lazy reads,
-    //      ~ms load even at Wikidata scale).
-    //   2. id_indices.bin.zst with KGLIIDX1 magic — 0.8.13 flat-CSR format
-    //      (eager decompress + HashMap rebuild; retained data fallback).
-    // Pre-0.14 bincode caches are ignored and rebuilt lazily.
-    let t = stage_timer();
-    if crate::graph::storage::GraphRead::is_disk(&graph.graph) {
-        if let Some(base) =
-            crate::graph::storage::disk::id_index::IdIndexBase::load_from(dir, &graph.interner)?
-        {
-            graph.id_indices = crate::graph::storage::disk::id_index::IdIndexStore::from_base(base);
-        } else {
-            let id_indices_path = dir.join("id_indices.bin.zst");
-            if id_indices_path.exists() {
-                if let Ok(compressed) = std::fs::read(&id_indices_path) {
-                    if let Ok(bytes) = zstd_decompress(&compressed) {
-                        if let Ok(Some(indices)) = read_id_indices_bin(&bytes, &graph.interner) {
-                            graph.id_indices.replace_with(indices);
-                        }
-                    }
-                }
-            }
-        }
-    }
-    log_stage("id_indices_load", t);
-
-    // 0.8.28+: `type_connectivity_cache` is populated lazily on first
-    // access (in `introspection/describe.rs`'s
-    // `compute_type_connectivity` fallback). Pre-loading it eagerly was
-    // costing 15+ s on slice-built graphs (128 M triples × 3 String
-    // allocations each) for data that most query workloads never touch.
-    // Read sites that miss the cache already degrade gracefully to a
-    // bounded edge scan.
-    //
-    // Opt-in eager load: `KGLITE_EAGER_TYPE_CONNECTIVITY=1`. Users that
-    // call `describe()` immediately after load can set this to amortize
-    // the cost into load instead of the first describe().
-    let t = stage_timer();
-    if std::env::var_os("KGLITE_EAGER_TYPE_CONNECTIVITY").is_some()
-        && !graph.has_type_connectivity_cache()
-    {
-        if let Ok(Some(triples)) = read_type_connectivity_bin(dir, &graph) {
-            if !triples.is_empty() {
-                *graph.type_connectivity_cache.write().unwrap() = Some(triples);
-            }
-        }
-    }
-    log_stage("type_connectivity_load", t);
-
+/// Load a disk graph's embeddings / timeseries / secondary-label sidecars.
+/// Split out of `load_disk_dir` to keep it under the function-complexity
+/// ceiling; cold load-time path with one shared fail-loud-on-corruption policy.
+fn load_disk_sidecars(dir: &std::path::Path, graph: &mut DirGraph) -> io::Result<()> {
     // Load embeddings if present. Absent file = fine (older graphs, or no
     // embeddings). A file that EXISTS but fails to decode is corruption and
     // must fail the load loudly — silently loading without embeddings would
@@ -2146,23 +2179,9 @@ fn load_disk_dir(dir: &std::path::Path) -> io::Result<Arc<DirGraph>> {
     // label degrade path (the reader returns Ok(false) when absent).
     // A present-but-undecodable file fails the load, same policy as
     // embeddings/timeseries above.
-    read_secondary_labels_bin(dir, &mut graph)
+    read_secondary_labels_bin(dir, graph)
         .map_err(|e| corrupt_sidecar_error("secondary_labels.bin.zst", &e))?;
-
-    // Backfill the connection_types O(1)-lookup cache from the loaded
-    // metadata. The v3 / file loader does this at line 1606 of read_v3;
-    // the disk loader was the only path that left it empty and relied
-    // on `has_connection_type`'s metadata-fallback branch. The fallback
-    // is correct on a freshly-loaded graph but flips into the wrong
-    // branch the moment any code path calls `register_connection_type`
-    // (which inserts into the cache and trips the "use cache" fast
-    // path on subsequent lookups). Backfilling here keeps the cache
-    // authoritative throughout the lifetime of the loaded graph.
-    graph.build_connection_types_cache();
-
-    log_stage("load_disk_dir_total", _load_t);
-
-    Ok(Arc::new(graph))
+    Ok(())
 }
 
 /// Read a disk graph's `metadata.json` and apply it to `graph`. A directory
